Added Earth Engine script for landcover classification
This is the code to find the location of an area using Google Earth Engine. Using this we locate the area , train the model to classify it to different land features like rice field , urbanized areas, water bodies and so on. This can be done with the help of QGIS. With the project we intend to find how a land changes over a period of time due to natural and other factors.